fix: fallback deps, bugs in import db on next10
