+++ src/artist/artist.module.ts
@@ -1,16 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
-import { Database } from '../../database';
+import { DatabaseModule } from 'src/database/database.module';
 
 @Module({
+  imports: [DatabaseModule],
   controllers: [ArtistController],
-  providers: [
-    ArtistService,
-    {
-      provide: 'Database',
-      useClass: Database,
-    },
-  ],
+  providers: [ArtistService],
 })
 export class ArtistModule {}
