+++ src/main.ts
@@ -1,6 +1,5 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { v4 } from 'uuid';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
