+++ src/user/user.module.ts
@@ -1,17 +1,12 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
-import { Database } from '../../database';
+import { DatabaseModule } from 'src/database/database.module';
 
 @Module({
+  imports: [DatabaseModule],
   controllers: [UserController],
-  providers: [
-    UserService,
-    {
-      provide: 'Database',
-      useClass: Database,
-    },
-  ],
+  providers: [UserService],
   exports: [UserService],
 })
 export class UserModule {}
